requirements.txt
google-cloud-aiplatform
vertexai
python-dotenv
gitpython
magika
streamlit